Implement authorinfo; split RevisionSlider out
